ModelsWrapper/requirements.txt
pickleshare==0.7.5
ktrain==0.39.0
deepl==1.16.1
numpy==1.21.6
translators==5.7.0
confluent-kafka==2.3.0
kafka-python==2.0.2
tensorflow==2.11.0